chore: remove unnecessary files and reorganize folders
