--- week-10-202402/comprandoKFC/comprandoKFC.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.9.34728.123
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "comprandoKFC", "comprandoKFC\comprandoKFC.vcxproj", "{2818B0ED-2BA3-44C2-90F7-A7595EFD6262}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2818B0ED-2BA3-44C2-90F7-A7595EFD6262}.Debug|x64.ActiveCfg = Debug|x64
-		{2818B0ED-2BA3-44C2-90F7-A7595EFD6262}.Debug|x64.Build.0 = Debug|x64
-		{2818B0ED-2BA3-44C2-90F7-A7595EFD6262}.Debug|x86.ActiveCfg = Debug|Win32
-		{2818B0ED-2BA3-44C2-90F7-A7595EFD6262}.Debug|x86.Build.0 = Debug|Win32
-		{2818B0ED-2BA3-44C2-90F7-A7595EFD6262}.Release|x64.ActiveCfg = Release|x64
-		{2818B0ED-2BA3-44C2-90F7-A7595EFD6262}.Release|x64.Build.0 = Release|x64
-		{2818B0ED-2BA3-44C2-90F7-A7595EFD6262}.Release|x86.ActiveCfg = Release|Win32
-		{2818B0ED-2BA3-44C2-90F7-A7595EFD6262}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {511BCA41-82B7-4CD7-ACDE-10A27A4C9FBD}
-	EndGlobalSection
-EndGlobal
