--- week-10-202402/comprandoKFC/comprandoKFC/x64/Debug/comprandoKFC.tlog/Cl.items.tlog
@@ -1,2 +1,0 @@
-C:\Users\marsr\source\repos\comprandoKFC\comprandoKFC\sukiWaterhouse.cpp;C:\Users\marsr\source\repos\comprandoKFC\comprandoKFC\x64\Debug\sukiWaterhouse.obj
-C:\Users\marsr\source\repos\comprandoKFC\comprandoKFC\x64\Debug\.NETFramework,Version=v4.7.2.AssemblyAttributes.cpp;C:\Users\marsr\source\repos\comprandoKFC\comprandoKFC\x64\Debug\.NETFramework,Version=v4.7.2.AssemblyAttributes.obj
