--- week-10-202402/comprandoKFC/comprandoKFC/x64/Debug/comprandoKFC.tlog/ResGen.command.1.tlog
@@ -1,4 +1,0 @@
-^C:\USERS\MARSR\SOURCE\REPOS\COMPRANDOKFC\COMPRANDOKFC\SUKIWATERHOUSE.RESX
-C:\Users\marsr\source\repos\comprandoKFC\comprandoKFC\comprandoKFC.vcxproj
-^C:\USERS\MARSR\SOURCE\REPOS\COMPRANDOKFC\COMPRANDOKFC\SUKIWATERHOUSE.RESX
-C:\Users\marsr\source\repos\comprandoKFC\comprandoKFC\comprandoKFC.vcxproj
