--- week-10-202402/comprandoKFC/comprandoKFC/x64/Debug/comprandoKFC.tlog/ResGen.read.1.tlog
@@ -1,4 +1,0 @@
-^C:\USERS\MARSR\SOURCE\REPOS\COMPRANDOKFC\COMPRANDOKFC\SUKIWATERHOUSE.RESX
-C:\USERS\MARSR\SOURCE\REPOS\COMPRANDOKFC\COMPRANDOKFC\COMPRANDOKFC.VCXPROJ
-^C:\USERS\MARSR\SOURCE\REPOS\COMPRANDOKFC\COMPRANDOKFC\SUKIWATERHOUSE.RESX
-C:\USERS\MARSR\SOURCE\REPOS\COMPRANDOKFC\COMPRANDOKFC\COMPRANDOKFC.VCXPROJ
