--- week-10-202402/comprandoKFC/comprandoKFC/x64/Debug/comprandoKFC.tlog/comprandoKFC.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.39.33519:TargetPlatformVersion=10.0.22621.0:TargetFrameworkVersion=v4.7.2::EnableManagedIncrementalBuild=true:
-Debug|x64|C:\Users\marsr\source\repos\comprandoKFC\|
