--- week-10-202402/comprandoKFC/comprandoKFC/x64/Debug/comprandoKFC.tlog/link.secondary.1.tlog
@@ -1,1 +1,0 @@
-^C:\USERS\MARSR\SOURCE\REPOS\COMPRANDOKFC\COMPRANDOKFC\X64\DEBUG\.NETFRAMEWORK,VERSION=V4.7.2.ASSEMBLYATTRIBUTES.OBJ|C:\USERS\MARSR\SOURCE\REPOS\COMPRANDOKFC\COMPRANDOKFC\X64\DEBUG\SUKIWATERHOUSE.OBJ
