--- week-10-202402/comprandoKFC/comprandoKFC/x64/Debug/comprandoKFC.tlog/metagen.read.1.tlog
@@ -1,1 +1,0 @@
-ÿþ
